refactor: changing mongo folder name
